thrust kernels for global stresses nearly complete
git-svn-id: svn+ssh://128.8.202.100/home/cdcl/mrfluids/fastmr@177 ce90a347-b419-4a64-a32e-38758aff188c

+++ particleSystem.h
@@ -70,7 +70,7 @@ public:
     float3 getWorldOrigin() { return m_params.worldOrigin; }
     float3 getCellSize() { return m_params.cellSize; }
 	
-	float getEdges();
+	uint getEdges();
 	int getGraphs();
 	
 	int getInteractionRadius() { return m_params.interactionr;}
